Ex2_js-basics-part2: Do task 3
Write a function that will count and display the number of odd
and even elements in the array to the console. If there is a zero
element in the array, then it is taken into account and displayed
separately. The returned value is an array of the form
[<number of even>, <number of odd>, <number of zeros>].
When performing the task, it is necessary to take into account that
the array can contain not only numbers, but also, for example,
signs, null, etc.